+++ packages/column-components/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file. The format
 is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this
 project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.1] - 2026-01-17
+
+- Modernized color picker component
+- Remove last `findDOMNode` usage for React 19 compatibility
+
 ## [1.5.0] - 2026-01-17
 
 Remove outdated `react-images` dependency and convert `PhotoViewer` component to