+++ packages/data-sheet/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file. The format
 is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this
 project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.2.4] - 2026-01-20
+
+- Modernize `react-color` dependency
+- Modernize data editor handling
+
 ## [2.2.3] - 2026-01-17
 
 - Fix issue with column resizing